Revert "Revert "fix: temp renamed app dir to prevent OOM""
This reverts commit 83a187bf11a4e72bd40e18357b3af139715b8e3f.
