+++ apps/web/_app/_trpc/HydrateClient.tsx
@@ -1,0 +1,8 @@
+"use client";
+
+import { createHydrateClient } from "_app/_trpc/createHydrateClient";
+import superjson from "superjson";
+
+export const HydrateClient = createHydrateClient({
+  transformer: superjson,
+});
