+++ apps/web/_app/future/(individual-page-wrapper)/apps/[slug]/layout.tsx
