+++ apps/web/_app/future/(individual-page-wrapper)/apps/installed/[category]/layout.tsx
