+++ apps/web/_app/future/(shared-page-wrapper)/(admin-layout)/settings/admin/impersonation/page.tsx
@@ -1,0 +1,10 @@
+import Page from "@pages/settings/admin/impersonation";
+import { _generateMetadata } from "_app/_utils";
+
+export const generateMetadata = async () =>
+  await _generateMetadata(
+    (t) => t("admin"),
+    (t) => t("impersonation")
+  );
+
+export default Page;
