+++ apps/web/_app/future/(shared-page-wrapper)/(admin-layout)/settings/admin/page.tsx
@@ -1,0 +1,10 @@
+import Page from "@pages/settings/admin/index";
+import { _generateMetadata } from "_app/_utils";
+
+export const generateMetadata = async () =>
+  await _generateMetadata(
+    () => "Admin",
+    () => "admin_description"
+  );
+
+export default Page;
