+++ apps/web/_app/future/(shared-page-wrapper)/(layout)/layout.tsx
