+++ apps/web/_app/future/(shared-page-wrapper)/(no-layout)/settings/admin/oAuth/oAuthView/page.tsx
@@ -1,0 +1,10 @@
+import Page from "@pages/settings/admin/oAuth/oAuthView";
+import { _generateMetadata } from "_app/_utils";
+
+export const generateMetadata = async () =>
+  await _generateMetadata(
+    () => "OAuth",
+    () => "Add new OAuth Clients"
+  );
+
+export default Page;
