--- apps/web/app/future/(shared-page-wrapper)/(admin-layout)/settings/admin/apps/page.tsx
@@ -1,10 +1,0 @@
-import Page from "@pages/settings/admin/apps/index";
-import { _generateMetadata } from "app/_utils";
-
-export const generateMetadata = async () =>
-  await _generateMetadata(
-    (t) => t("apps"),
-    (t) => t("admin_apps_description")
-  );
-
-export default Page;
