--- apps/web/app/future/(shared-page-wrapper)/(admin-layout)/settings/admin/flags/page.tsx
@@ -1,10 +1,0 @@
-import Page from "@pages/settings/admin/flags";
-import { _generateMetadata } from "app/_utils";
-
-export const generateMetadata = async () =>
-  await _generateMetadata(
-    () => "Feature Flags",
-    () => "Here you can toggle your Cal.com instance features."
-  );
-
-export default Page;
