--- apps/web/app/future/(shared-page-wrapper)/(layout)/event-types/page.tsx
@@ -1,10 +1,0 @@
-import EventTypes from "@pages/event-types";
-import { _generateMetadata } from "app/_utils";
-
-export const generateMetadata = async () =>
-  await _generateMetadata(
-    (t) => t("event_types_page_title"),
-    (t) => t("event_types_page_subtitle")
-  );
-
-export default EventTypes;
