--- apps/web/app/future/(shared-page-wrapper)/(settings-layout)/settings/admin/organizations/page.tsx
@@ -1,11 +1,0 @@
-import { _generateMetadata } from "app/_utils";
-
-import Page from "@calcom/features/ee/organizations/pages/settings/admin/AdminOrgPage";
-
-export const generateMetadata = async () =>
-  await _generateMetadata(
-    (t) => t("organizations"),
-    (t) => t("orgs_page_description")
-  );
-
-export default Page;
