--- apps/web/app/future/(shared-page-wrapper)/(settings-layout)/settings/admin/users/add/page.tsx
@@ -1,11 +1,0 @@
-import { _generateMetadata } from "app/_utils";
-
-import Page from "@calcom/features/ee/users/pages/users-add-view";
-
-export const generateMetadata = async () =>
-  await _generateMetadata(
-    () => "Add new user",
-    () => "Here you can add a new user."
-  );
-
-export default Page;
